requirements.txt
Django==1.5.1
South==0.8.1
distribute==0.6.24
dj-database-url==0.2.1
django-annoying==0.7.7
googlemaps==1.0.2
gunicorn==0.17.4
psycopg2==2.5
wsgiref==0.1.2
